requirements.txt
Flask
flask_cors
requests
GitPython
